+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Bugfixes
 
  - Remove redundant alembic upgrade that can cause upgrade errors
- - Fix moving Conditional code during upgrade
+ - Fix moving Conditional/input code during upgrade
+ - Generate Conditional/input code for next upgrade
  - Fix MQTT Input ([#685](https://github.com/kizniche/mycodo/issues/685))
 
 # TODO Add db revision that saves conditional code to user code directory